Refactor map feature and improve UI handling
