+++ lib/core/utils/string_extensions.dart
@@ -1,0 +1,11 @@
+extension NullableStringUtils on String? {
+  String? emptyToNull() {
+    final value = this;
+    if (value == null) return null;
+    final trimmed = value.trim();
+    if (trimmed.isEmpty) {
+      return null;
+    }
+    return trimmed;
+  }
+}
