+++ lib/features/geofence/models/geofence_result.dart
@@ -1,0 +1,9 @@
+class GeofenceResult {
+  const GeofenceResult({
+    required this.inside,
+    required this.statusMessage,
+  });
+
+  final bool inside;
+  final String statusMessage;
+}
