+++ lib/features/geofence/presentation/widgets/status_card.dart
@@ -1,0 +1,51 @@
+import 'package:flutter/material.dart';
+
+class StatusCard extends StatelessWidget {
+  const StatusCard({
+    super.key,
+    required this.accuracyText,
+    required this.statusMessage,
+    this.errorMessage,
+  });
+
+  final String accuracyText;
+  final String statusMessage;
+  final String? errorMessage;
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      margin: EdgeInsets.zero,
+      color: Colors.white.withOpacity(0.95),
+      elevation: 6,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+      child: Padding(
+        padding: const EdgeInsets.all(12),
+        child: ConstrainedBox(
+          constraints: const BoxConstraints(maxWidth: 320),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              const Text(
+                'জিপিএস এর অবস্থা:',
+                style: TextStyle(fontWeight: FontWeight.bold),
+              ),
+              const SizedBox(height: 6),
+              Text('সঠিকতা: $accuracyText'),
+              const SizedBox(height: 4),
+              Text(statusMessage),
+              if (errorMessage != null) ...[
+                const SizedBox(height: 6),
+                Text(
+                  errorMessage!,
+                  style: const TextStyle(color: Colors.red),
+                ),
+              ],
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
