+++ pubspec.yaml
@@ -39,6 +39,7 @@ dependencies:
   geolocator: ^12.0.0
   shared_preferences: ^2.3.2
   image_picker: ^1.1.1
+  provider: ^6.1.2
 
 dev_dependencies:
   flutter_launcher_icons: ^0.13.1
